src/app/requirements.txt
gradio>=4.0
databricks-sql-connector>=3.0
pandas
